+++ src/cli/commands/iot-download-bulk.ts
@@ -32,7 +32,7 @@ export default (program: CommanderStatic) => {
         .option("-p, --passkey <passkey>", `passkey`)
         .option("-y, --retry <number>", "retry attempts before giving up", "3")
         .option("-v, --verbose", "verbose output")
-        .description(`${color("download the timeseries from mindsphere")}`)
+        .description(`${color("download the timeseries data in bulk from mindsphere *")}`)
         .action((options) => {
             (async () => {
                 try {
@@ -102,12 +102,12 @@ function checkParameters(options: any) {
     !options.dir &&
         errorLog("Missing dir name for download-bulk command. Run mc db --help for full syntax and examples.", true);
 
-    !options.assetid && errorLog(" You have to specify assetid. Run  mc df --help for full syntax and examples.", true);
+    !options.assetid && errorLog(" You have to specify assetid. Run  mc db --help for full syntax and examples.", true);
 
     !options.aspectname &&
-        errorLog(" You have to specify aspectname. Run  mc df --help for full syntax and examples.", true);
+        errorLog(" You have to specify aspectname. Run  mc db --help for full syntax and examples.", true);
 
-    !options.from && errorLog(" You have to specify from date. Run  mc df --help for full syntax and examples.", true);
+    !options.from && errorLog(" You have to specify from date. Run  mc db --help for full syntax and examples.", true);
 
-    !options.to && errorLog(" You have to specify to date. Run  mc df --help for full syntax and examples.", true);
+    !options.to && errorLog(" You have to specify to date. Run  mc db --help for full syntax and examples.", true);
 }